перенёс удаление пользователя в @after

+++ src/test/java/ru/praktikum/tests/RegisterTest.java
@@ -7,12 +7,13 @@ import ru.praktikum.pages.RegisterPage;
 import ru.praktikum.utils.RandomData;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import ru.praktikum.utils.UserApi;
-
-import java.time.Duration;
+import org.junit.After;
 
 import static org.junit.Assert.assertTrue;
 
 public class RegisterTest extends BaseTest {
+
+    private String accessToken;
 
     @Test
     public void successfulRegistrationTest() {
@@ -68,7 +69,7 @@ public class RegisterTest extends BaseTest {
         String email = RandomData.generateEmail();
         String password = "123456";
 
-        String accessToken = UserApi.createUser(email, password, "Test User");
+        accessToken = UserApi.createUser(email, password, "Test User");
 
         MainPage mainPage = new MainPage(driver);
         mainPage.clickLoginButton();
@@ -77,8 +78,6 @@ public class RegisterTest extends BaseTest {
         loginPage.login(email, password);
 
         assertTrue(driver.getPageSource().contains("Личный Кабинет"));
-
-        UserApi.deleteUser(accessToken);
     }
 
     @Test
@@ -87,7 +86,7 @@ public class RegisterTest extends BaseTest {
         String email = RandomData.generateEmail();
         String password = "123456";
 
-        String accessToken = UserApi.createUser(email, password, "Test User");
+        accessToken = UserApi.createUser(email, password, "Test User");
 
         MainPage mainPage = new MainPage(driver);
         mainPage.clickPersonalAccount();
@@ -96,8 +95,6 @@ public class RegisterTest extends BaseTest {
         loginPage.login(email, password);
 
         assertTrue(driver.getPageSource().contains("Личный Кабинет"));
-
-        UserApi.deleteUser(accessToken);
     }
 
     @Test
@@ -106,7 +103,7 @@ public class RegisterTest extends BaseTest {
         String email = RandomData.generateEmail();
         String password = "123456";
 
-        String accessToken = UserApi.createUser(email, password, "Test User");
+        accessToken = UserApi.createUser(email, password, "Test User");
 
         MainPage mainPage = new MainPage(driver);
         mainPage.clickLoginButton();
@@ -119,7 +116,12 @@ public class RegisterTest extends BaseTest {
         loginPage.login(email, password);
 
         assertTrue(driver.getPageSource().contains("Личный Кабинет"));
+    }
 
-        UserApi.deleteUser(accessToken);
+    @After
+    public void deleteUser() {
+        if (accessToken != null) {
+            UserApi.deleteUser(accessToken);
+        }
     }
 }
